requirements.txt
Django==5.0.3
django-filter==21.1
drf-yasg==1.21.5
flake8==7.0.0
flake8-absolute-import==1.0.0.2
flake8-annotations==3.0.1
flake8-broken-line==1.0.0
flake8-bugbear==24.1.17
flake8-eradicate==1.5.0
flake8-mutable==1.2.0
flake8-pep3101==2.1.0
flake8-plugin-utils==1.3.3
flake8-polyfill==1.0.2
flake8-print==5.0.0
flake8-printf-formatting==1.1.2
flake8-pytest-style==1.7.2
flake8-quotes==3.3.2
flake8-tuple-unpack-limit==0.0.1
openpyxl==3.1.3
pandas==2.2.2
pillow==10.2.0
psycopg2-binary==2.9.9
python-docx==1.1.2
requests==2.31.0
ruff==0.4.8
sqlparse==0.4.4
xlrd==2.0.1
